requirements.txt
flask
flask-cors
pymongo
python-dotenv
requests
gunicorn